+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+* Added support for parsing `import.meta` in JavaScript modules.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.19] - 2018-03-30